refactor: now files more organise and crimes against humanity no longer exist
