+++ compile.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-kotlinc src/*.kt  -cp "kotlinx-cli-jvm-0.3.6.jar" -include-runtime -d app.jar
+kotlinc $(find src -name "*.kt") -cp "kotlinx-cli-jvm-0.3.6.jar" -include-runtime -d app.jar
