+++ src/models/Action.kt
@@ -1,0 +1,2 @@
+package models
+enum class Action { READ, WRITE, EXECUTE }
