+++ src/models/UserData.kt
@@ -1,0 +1,2 @@
+package models
+data class UserData(val salt: String, val hash: String)
